+++ src/contexts/kiosk/useKiosk.ts
@@ -1,0 +1,10 @@
+import { useContext } from 'react';
+import { KioskContext } from './KioskContext';
+
+export const useKiosk = () => {
+  const context = useContext(KioskContext);
+  if (!context) {
+    throw new Error('useKiosk must be used within a KioskProvider');
+  }
+  return context;
+};
